HW-14/pyproject.toml
[tool.poetry]
name = "hw-14"
version = "0.1.0"
description = ""
authors = ["Ksushondrik <ksushondrik369@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.112.0"
alembic = "^1.13.2"
sqlalchemy = "^2.0.32"
asyncpg = "^0.29.0"
uvicorn = "^0.30.6"
pydantic = {extras = ["email"], version = "^2.8.2"}
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
fastapi-users = {extras = ["sqlalchemy"], version = "^13.0.0"}
libgravatar = "^1.0.4"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-multipart = "^0.0.9"
fastapi-mail = "^1.4.1"
python-dotenv = "^1.0.1"
redis = "^5.0.8"
fastapi-limiter = "^0.1.6"
cloudinary = "^1.41.0"
pytest = "^8.3.2"
pytest-cov = "^5.0.0"


[tool.poetry.group.dev.dependencies]
sphinx = "^8.0.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
addopts = "--doctest-modules"
testpaths = ["tests",]
pythonpath = "."
